Separeted the ReportTable Component into 2 tider components
